Split jiterpreter trace generation out of jiterpreter.ts into a new file
